refactor: Header types 분리
