+++ src/components/Header/themes.ts
@@ -1,15 +1,4 @@
-
-export interface Theme {
-  label: string;
-  title: string;
-  description: string;
-  backgroundColor: string;
-  imageSrc: string;
-}
-
-export interface Themes {
-  [key: string]: Theme;
-}
+import type { Themes } from '@/components/Header/types';
 
 export const themes: Themes = {
   theme1: {
