+++ src/components/Header/types.ts
@@ -1,0 +1,12 @@
+export interface Theme {
+    label: string;
+    title: string;
+    description: string;
+    backgroundColor: string;
+    imageSrc: string;
+  }
+  
+  export interface Themes {
+    [key: string]: Theme;
+  }
+
